VERONICA/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
pinecone-client==3.0.0
pypdf==3.17.4
python-docx==1.1.0
chromadb==0.4.22
ollama==0.1.6
sentence-transformers==2.2.2
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
